[Solana]: Minor changes
