+++ rust/chains/tw_solana/src/modules/instruction_builder/mod.rs
@@ -3,25 +3,20 @@
 // Copyright © 2017 Trust Wallet.
 
 use crate::address::SolanaAddress;
-use crate::blockhash::Blockhash;
 use crate::instruction::Instruction;
-use crate::program::stake_program::StakeProgram;
-use tw_coin_entry::error::SigningResult;
 
 pub mod stake_instruction;
 pub mod system_instruction;
 pub mod token_instruction;
 
-use stake_instruction::{Authorized, Lockup, StakeInstructionBuilder};
 use system_instruction::SystemInstructionBuilder;
 
-// TODO rename to `InstructionBuilder`.
 #[derive(Default)]
-pub struct InstructionBuilder1 {
+pub struct InstructionBuilder {
     instructions: Vec<Instruction>,
 }
 
-impl InstructionBuilder1 {
+impl InstructionBuilder {
     /// Adds an `advance_nonce` instruction if the nonce account provided.
     pub fn maybe_advance_nonce(
         &mut self,
@@ -61,47 +56,5 @@ impl InstructionBuilder1 {
 
     pub fn output(self) -> Vec<Instruction> {
         self.instructions
-    }
-}
-
-pub struct DepositStakeArgs {
-    pub sender: SolanaAddress,
-    pub validator: SolanaAddress,
-    pub stake_account: Option<SolanaAddress>,
-    pub recent_blockhash: Blockhash,
-    pub lamports: u64,
-    pub space: u64,
-}
-
-/// Consists of complex instruction builders.
-pub struct InstructionBuilder;
-
-impl InstructionBuilder {
-    // TODO move to `StakeInstructionBuilder`.
-    pub fn deposit_stake(args: DepositStakeArgs) -> SigningResult<Vec<Instruction>> {
-        let stake_addr = args.stake_account.unwrap_or_else(|| {
-            // no stake address specified, generate a new unique
-            StakeProgram::address_from_recent_blockhash(args.sender, args.recent_blockhash)
-        });
-        let seed = args.recent_blockhash.to_string();
-
-        let authorized = Authorized {
-            staker: args.sender,
-            withdrawer: args.sender,
-        };
-        let lockup = Lockup::default();
-
-        Ok(vec![
-            SystemInstructionBuilder::account_with_seed(
-                args.sender,
-                stake_addr,
-                args.sender,
-                seed,
-                args.lamports,
-                args.space,
-            ),
-            StakeInstructionBuilder::stake_initialize(stake_addr, authorized, lockup),
-            StakeInstructionBuilder::delegate(stake_addr, args.validator, args.sender),
-        ])
     }
 }
